+++ src/app/app.module.ts
@@ -1,4 +1,4 @@
-import {DEFAULT_CURRENCY_CODE, NgModule} from '@angular/core';
+import {DEFAULT_CURRENCY_CODE, InjectionToken, NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {AppRoutingModule} from './app-routing.module';
@@ -17,6 +17,13 @@ import {MatButtonModule} from "@angular/material/button";
 import {MatListModule} from "@angular/material/list";
 import {OrderModule} from "./order/order.module";
 import {AdminModule} from "./admin/admin.module";
+import {HttpClientModule} from "@angular/common/http";
+import {environment} from "../environments/environment";
+import {Environment} from "./core/models/environment";
+
+export const ENV_TOKEN = new InjectionToken<Environment>(
+  'app.environment.token'
+);
 
 @NgModule({
   declarations: [
@@ -33,6 +40,7 @@ import {AdminModule} from "./admin/admin.module";
     MatSidenavModule,
     MatButtonModule,
     MatListModule,
+    HttpClientModule,
     AdminModule,
     AppRoutingModule,
   ],
@@ -41,6 +49,7 @@ import {AdminModule} from "./admin/admin.module";
     {provide: APP_CONFIG, useValue: APP_DI_CONFIG},
     {provide: SESSION_ID, useFactory: GeneratorFactory(20), deps: [GeneratorService]},
     {provide: DEFAULT_CURRENCY_CODE, useValue: 'USD'},
+    {provide: ENV_TOKEN, useFactory: () => environment},
   ],
   bootstrap: [AppComponent]
 })