+++ src/app/shared/guards/product.resolver.ts
@@ -2,23 +2,17 @@ import {ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot} from "@ang
 import {ProductModel} from "../../product/models/product.model";
 import {ProductRepository} from "../../product/services/product.repository";
 import {Injectable} from "@angular/core";
+import {Observable} from "rxjs";
 
 export function resolveProductWithRedirect(redirectUrl: string) {
-  return { data: { redirectUrl }, resolve: { product: ProductResolver }};
+  return {data: {redirectUrl}, resolve: {product: ProductResolver}};
 }
 
 @Injectable({providedIn: 'root'})
-export class ProductResolver implements Resolve<ProductModel | null> {
-  constructor(private productRepository: ProductRepository, private router: Router) {}
+export class ProductResolver implements Resolve<ProductModel> {
+  constructor(private productRepository: ProductRepository) {}
 
-  async resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<ProductModel | null> {
-    const product = await this.productRepository.getProduct(+route.params['productID']);
-    if (product) {
-      return product;
-    }
-
-    this.router.navigate([route.data.redirectUrl]);
-
-    return null;
+  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<ProductModel> {
+    return this.productRepository.getProduct(+route.params['productID']);
   }
 }